requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
sentence-transformers
